colocando el audio

+++ js/behaivor.js
@@ -2,7 +2,7 @@
 
 var Comportamiento=function(game){
     this.game = game;
-     
+    
 }
 
 
@@ -16,33 +16,39 @@ Comportamiento.prototype={
         this.walls.enableBody = true;
         
         for (i = 0; i < this.WALLZISE; i++){
-                this.wallA.push(this.walls.create(0, 0+60*i, 'killerwall'));
+            this.wallA.push(this.walls.create(0, 0+60*i, 'killerwall'));
         }
         for (i = 0; i < this.WALLZISE; i++){
-                this.wallB.push(this.walls.create(710, 0+60*i, 'killerwall'));
+            this.wallB.push(this.walls.create(710, 0+60*i, 'killerwall'));
         }
         this.objects = this.game.add.group();
         this.objects.enableBody = true;
         this.startcrashing = false;
         this.timestart = 0;
+        this.active = false;
     },
     
     update: function(){
-        for (i = 0; i < this.WALLZISE; i++){
-            this.wallA[i].body.velocity.x = WallVelocity;
+        if(this.active){
+            for (i = 0; i < this.WALLZISE; i++){
+                this.wallA[i].body.velocity.x = WallVelocity;
+            }
+            for (i = 0; i < this.WALLZISE; i++){
+                this.wallB[i].body.velocity.x = -WallVelocity;
+            }
+                   
+            //this.game.physics.arcade.overlap(this.player, this.walls, this.crash());
+            this.game.physics.arcade.collide(this.player, this.walls, this.crash,null, this);
+        
         }
-        for (i = 0; i < this.WALLZISE; i++){
-            this.wallB[i].body.velocity.x = -WallVelocity;
-        }
-            
-            
-             //this.game.physics.arcade.overlap(this.player, this.walls, this.crash());
-         this.game.physics.arcade.collide(this.player, this.walls, this.crash,null, this);
+    },
+    setActive :function(value){
+        this.active=value;
     },
     
     crash : function(){
-                //console.log("xD");
-                
+        if(this.active){
+                //console.log("xD");        
                 
                 //alert(this.objects);
                 this.blood = this.objects.create(this.player.body.x+10,this.player.body.y+10, 'blood');
@@ -64,11 +70,14 @@ Comportamiento.prototype={
                         
                     }
                 }
+          }
     },
     
     destroy_all: function(){
-        this.walls.destroy();
-        this.objects.destroy();
+        if(this.active){
+            this.walls.destroy();
+            this.objects.destroy();
+        }
     } 
     
 };
